refactor: move gh package to internal
