+++ cmd/run/run.go
@@ -11,9 +11,9 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 
+	"github.com/aweris/gale/internal/gh"
 	"github.com/aweris/gale/pkg/config"
 	"github.com/aweris/gale/pkg/gale"
-	"github.com/aweris/gale/pkg/gh"
 	"github.com/aweris/gale/pkg/model"
 	"github.com/aweris/gale/pkg/repository"
 )
