+++ pkg/gale/gale.go
@@ -12,8 +12,8 @@ import (
 	"github.com/aweris/gale/internal/dagger/binaries"
 	"github.com/aweris/gale/internal/dagger/images"
 	"github.com/aweris/gale/internal/dagger/services"
+	"github.com/aweris/gale/internal/gh"
 	"github.com/aweris/gale/pkg/config"
-	"github.com/aweris/gale/pkg/gh"
 	"github.com/aweris/gale/pkg/model"
 )
